pyproject.toml
[project]
name = "ajudanteimpressaov2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.0",
    "Pillow>=10.0",
    "PySide6>=6.7",
]
